rename vqr module to cvqf ahead of refactor
