+++ src/vqr/__init__.py
@@ -1,5 +1,5 @@
 # isort: skip_file
-from vqr.vqr import QuantileFunction
+from vqr.cvqf import QuantileFunction
 from vqr.solvers import VQRSolver, VQRSolution
 from vqr.api import VectorQuantileEstimator, VectorQuantileRegressor
 from pkg_resources import DistributionNotFound, get_distribution
