+++ src/vqr/cvqf.py
